+++ core/Tracker/Action.php
@@ -12,10 +12,8 @@ namespace Piwik\Tracker;
 use Exception;
 use Piwik\Common;
 use Piwik\Container\StaticContainer;
-use Piwik\Piwik;
 use Piwik\Plugin\Dimension\ActionDimension;
 use Piwik\Plugin\Manager;
-use Piwik\Tracker;
 use Psr\Log\LoggerInterface;
 
 /**