+++ core/Updates/2.14.0-b2.php
@@ -11,7 +11,6 @@ namespace Piwik\Updates;
 
 use Piwik\Updater;
 use Piwik\Updates;
-use Piwik\Db;
 use Piwik\Updater\Migration\Factory as MigrationFactory;
 
 class Updates_2_14_0_b2 extends Updates