+++ core/Updates/4.0.0-b1.php
@@ -13,7 +13,6 @@ use Piwik\Date;
 use Piwik\Plugins\UsersManager\Model;
 use Piwik\Updater;
 use Piwik\Updates as PiwikUpdates;
-use Piwik\Updater\Migration;
 use Piwik\Updater\Migration\Factory as MigrationFactory;
 
 /**