+++ plugins/Goals/tests/Integration/APITest.php
@@ -34,9 +34,6 @@ class APITest extends IntegrationTestCase
         parent::setUp();
         $this->api = API::getInstance();
 
-        Fixture::createAccessInstance();
-        Piwik::setUserHasSuperUserAccess();
-
         Fixture::createWebsite('2014-01-01 00:00:00');
         Fixture::createWebsite('2014-01-01 00:00:00');
     }